[docs] Rename theme-default to default-theme
