+++ docs/src/pages/customization/default-theme/DefaultTheme.js
