+++ docs/src/pages/layout/css-in-js/css-in-js.md
@@ -4,7 +4,7 @@
 
 Sometimes, wrapping things inside a layout component doesn't make much sense when
 everything you need can be handled by a small CSS change. By using the
-[`breakpoints`](/customization/theme-default?expend-path=$.breakpoints) attribute of the theme, you can utilise the same breakpoints used
+[`breakpoints`](/customization/default-theme?expend-path=$.breakpoints) attribute of the theme, you can utilise the same breakpoints used
 for the [Grid](/layout/grid) and [Hidden](/layout/hidden) components directly in your component.
 
 This can be accomplished using [CSS-in-JS](/customization/css-in-js).
