+++ docs/src/pages/style/typography/typography.md
@@ -44,6 +44,6 @@ Material-UI default typography configuration only relies on 300, 400 and 500 fon
 ## CSS in JS
 
 In some situation you might not be able to use the `Typography` component.
-Hopefully, you might be able to take advantage of the [`typography`](/customization/theme-default?expend-path=$.typography) keys of the theme.
+Hopefully, you might be able to take advantage of the [`typography`](/customization/default-theme?expend-path=$.typography) keys of the theme.
 
 {{"demo": "pages/style/typography/TypographyTheme.js"}}
